[front] chore: Split lib/api/projects
